Refactored to wrap and export a more complex component.
The component being exported is now called micro-frontend. It
is made with a series of tabs that use routing, both eager and
lazy and also a tab which lazily loads its content without using
the router.

This component is wrapped by wrapper component, which shims the
router to keep it internal and provides a new API for selecting
tabs.